Les dégâts sont affichés en double.
+ Le bonus d'except est calculé dans le baseWeapon, à la place de dans
le CombatStrategy.

+++ Scripts/Engines/Combat/StrategiesDistance.cs
@@ -54,7 +54,7 @@ namespace Server.Engines.Combat
 
         public override SkillName ToucherSkill { get { return SkillName.ArmeDistance; } }
 
-        protected override double ComputerDegats(Mobile atk, int basedmg, bool skillup)
+        protected override double ComputerDegats(Mobile atk, double basedmg, bool skillup)
         {
             if (skillup)
                 CheckSkillGain(atk, SkillName.Anatomie);
